Working OpenF1 API
Got the OpenF1 API working. Currently working out how data for specific sessions will be gathered, but it should be moving smoothly. As it stands with this variation, there are two new dataframes, that being sessions_data, and race_sessions_2025. Moving forwards these will be modified and updated, plus many more dataframes are going to be created.