Restarting and Improving Data Cleaning
There are some major flaws with the pervious data cleaning as it was not effectively planned and checked, so I am taking an hour or so to redo this code in order to isolate the the proper data. Notably, two drivers names were present that should not have been, while others were not, so I am more methodically planning out these sections now that I know that I want to visualize the data that way.